Feed NLQ Data
Added a new class that provide the QA dataset in batch form. This will
serve as a replacement to RelationEntityBatcher which works with
triplets. QuestionBatcher on the other hand works with Natural Language
Questions. Transformer.py now serves as an example file on how to work
with this batcher.

Update